Refactor sturcture of pipeline_executor

+++ modyn/config/schema/pipeline.py
@@ -317,7 +317,7 @@ class TrainingConfig(BaseModel):
     dataloader_workers: int = Field(
         description="The number of data loader workers on the trainer node that fetch data from storage.", ge=1
     )
-    batch_size: float = Field(description="The batch size to be used during training.", ge=1)
+    batch_size: int = Field(description="The batch size to be used during training.", ge=1)
     use_previous_model: bool = Field(
         description=(
             "If True, on trigger, we continue training on the model outputted by the previous trigger. If False, "
